fixed navbar and routes
