--- client/src/components/Nav/style.css
@@ -1,28 +1,0 @@
-.nav-style{
-    background-color: crimson;
-    color: white;
-}
-
-@media (max-width: 768px) {  
-  
-   #mobileFlex {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    }
-}
-
-/*
- * Made by Erik Terwan
- * 24th of November 2015
- * MIT License
- *
- *
- * If you are thinking of using this in
- * production code, beware of the browser
- * prefixes.
- */
-
-
- 
-
