+++ client/src/index.css
@@ -11,3 +11,28 @@ code {
   font-family: source-code-pro, Menlo, Monaco, Consolas, 'Courier New',
     monospace;
 }
+
+ /* NAV */
+
+ .nav-style {
+  background-color: crimson;
+  color: white;
+  /* margin: 5px; */
+}
+.nav-button{
+  color: white !important;
+}
+.linkText {
+  text-decoration: none;
+  text-transform: uppercase;
+  color: white;
+  margin: 5px;
+}
+@media (max-width: 768px) {  
+
+ #mobileFlex {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  }
+}
